Apply backface culling based on triangle winding order
To correctly backface cull triangles we either need to:
 - compute a face normal from vertex normals and then do a dot
   product with view direction.
 - Or: detected the winding order of the triangle post-projection.

The first method is slow and does not work for meshes that do not
provide normals (or, at least, we would need to compute normals based on
vertices and winding order).
This commit implements the second method, which tends to be faster and I
believe it's how it is typically done in hardware.